ISIS-2441: separate and encapsulate change tracking concerns
