+++ core/runtime/src/main/java/org/apache/isis/core/runtime/persistence/changetracking/AuditEntry.java
@@ -16,13 +16,15 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.core.runtime.persistence.transaction;
+package org.apache.isis.core.runtime.persistence.changetracking;
 
-import java.util.Set;
+import lombok.NonNull;
+import lombok.Value;
 
-//TODO[ISIS-2441] intermediate for refactoring keep?
-public interface HasEnlistedForAuditing {
+@Value(staticConstructor = "of")
+final class AuditEntry {
 
-    Set<AuditEntry> getChangedObjectProperties();
-
+    @NonNull private final AdapterAndProperty adapterAndProperty;
+    @NonNull private final PreAndPostValues preAndPostValues;
+    
 }
