+++ core/runtime/src/main/java/org/apache/isis/core/runtime/persistence/changetracking/HasEnlistedForMetrics.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.isis.core.runtime.persistence.transaction;
+package org.apache.isis.core.runtime.persistence.changetracking;
 
 //TODO[ISIS-2441] intermediate for refactoring keep?
 public interface HasEnlistedForMetrics {
